Enable using weights to plot FES
Allows passing an array of weights (for each conformation in a trajectory) to be reflected in building an FES (histogram). This enables usage of molPX to visualise biased trajectories, eg. obtained with metadynamics or umbrella sampling.
For metadynamics, weights can be obtained using the algorithm proposed by Tiwary and Parrinello JPCB 2014 (doi:10.1021/jp504920s).